feat: added `run` subcommand for running models commands with rendered configs
Sometimes you want to run a command for each configuration instance that
depends on one ore more configuration parameters. For example, if you
have a model that only writes to standard output, then you will have to
direct the output to a file with `>`. But if you have multiple
configurations you will want to send the output of each command to
different file.

The run command will load a configuration and run a set of shell
commands that have been rendered using the config tree as a context.